more detailed tests of what's allowed and where (#450)
